Update and rename coulomb_law.py to coulombs_law.py
